🟡(change): type definitions
