+++ internal/api/services/types/table_change_alert.go
@@ -1,0 +1,11 @@
+package types
+
+type TopicName struct {
+	Value string `json:"value"`
+	Label string `json:"label"`
+}
+
+type TableName struct {
+	Value string `json:"value"`
+	Label string `json:"label"`
+}
